requirements.txt
yaml2jsonnet
deepdiff
bson
benedict
pandas
glob2
collections